Decompose infrastructure stacks and change DynamoDB schema
The original design did not account for emitting stacks in regions
other than the region in which the infrastructure was deployed.

This is fixed by
* Separating the Lambda function IAM role into it's own stack
* Moving to a model where the "consumer" stack which contains the
  SNS topic and Lambda function are deployed in each region that
  you want to support receiving emissions in

The DynamoDB schema was also changed to prevent multiple stacks
in the same account from overwriting each other. Originally this
was intentional in the design but after working through trying to
use it, it became apparent that it wasn't ideal.

The new schema continues to use the account id as the partition key
but changes the sort key from "category" to the GUID of the
CloudFormation stack that emitted the record.

Additional cleanup includes
* Deleting the ServiceToken attribute so it's not stored in the DyanmoDB
  record
* Adding an attribute with the region of the emitting stack
* Removing the table provsioning from the Lambda function as this is
  handled in the dedicated DynamoDB CloudFormation template
* Added documentation and examples of how to constrain emitting stacks
  to only regions in which you've deployed the consuming infrastructure